+++ komposable-architecture-test/src/main/java/com/toggl/komposable/test/utils/JavaLogger.kt
@@ -1,4 +1,4 @@
-package com.toggl.komposable.test.store
+package com.toggl.komposable.test.utils
 
 class JavaLogger(private val logger: java.util.logging.Logger) : Logger {
     override fun log(level: LogLevel, message: String?, error: Throwable?) {
